+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Changelog for NeoFS Node
 - "new transaction callback finished with error" logs for "already exists in mempool" error (#3218)
 - Panic if logger is not set in FSTree (#3220)
 - Not initialized FSTree in the write-cache (#3220)
+- Potential approval of network map candidate with duplicated attribute by IR (#3224)
 
 ### Changed